More generic plugin mechanism, and call importers from button.
Since we cannot load assemblies at runtime on monotouch, we have to have a reference to classes inside the assemblies to load plugins. The solution (at least so far) is to have a simple class in the UI project that registers all available plugins.

Importer-selection for multiple possibilities is not implemented.